requirements.txt
python-nmap
colorama
tabulate
requests